(1.0/0.0) ei kelpaa Mac-ympÃ¤ristÃ¶ssÃ¤... MitenkÃ¤s cast doubleksi?

+++ potential.h
@@ -1,12 +1,12 @@
 /*
- * potential.h $Id: potential.h,v 1.32 2006-11-09 16:45:12 jatoivol Exp $
+ * potential.h $Id: potential.h,v 1.33 2006-11-09 16:55:04 jatoivol Exp $
  */
 
 #ifndef __POTENTIAL_H__
 #define __POTENTIAL_H__
 
 #ifndef HUGE_VAL
-#define HUGE_VAL (1.0/0.0)
+#define HUGE_VAL (double)(1.0/0.0)
 #endif
 
 typedef struct pot_array_t {
